마법사 구현중
마법사 4번째 스킬 상태이상 시스템 최적화 중

+++ Player.cs
@@ -6,8 +6,8 @@ using static System.Console;
 
 public abstract class Player : Character
 {
-    public int MaxMp { get; set; }
-    public int Mp { get; set; }
+    public double MaxMp { get; set; }
+    public double Mp { get; set; }
     public List<Skill> skills = new();
 
     public void PrintSkills()
